Move RenderMixin to extras.models.mixins, Rename RenderMixin to RenderTemplateMixin
